refactor(navigation): migrate navigation to manager module
