+++ .idea/gradle.xml
@@ -26,7 +26,6 @@
             <option value="$PROJECT_DIR$/core/database" />
             <option value="$PROJECT_DIR$/core/datasource" />
             <option value="$PROJECT_DIR$/core/design-system" />
-            <option value="$PROJECT_DIR$/core/dispatcher" />
             <option value="$PROJECT_DIR$/core/manager" />
             <option value="$PROJECT_DIR$/core/repository" />
             <option value="$PROJECT_DIR$/core/resources" />
