+++ application/src/androidMain/kotlin/com/neoutils/neoregex/NeoAndroidApp.kt
@@ -22,7 +22,6 @@ import android.app.Application
 import android.content.Context
 import com.neoutils.neoregex.core.database.di.databaseModule
 import com.neoutils.neoregex.core.datasource.di.dataSourceModule
-import com.neoutils.neoregex.core.dispatcher.di.dispatcherModule
 import com.neoutils.neoregex.core.manager.di.managerModule
 import com.neoutils.neoregex.core.repository.di.repositoryModule
 import com.neoutils.neoregex.feature.matcher.di.matcherModule
@@ -47,7 +46,6 @@ class NeoAndroidApp : Application() {
                 dataSourceModule,
                 databaseModule,
                 repositoryModule,
-                dispatcherModule,
                 matcherModule,
                 validatorModule,
                 savedModule,
