+++ application/src/desktopMain/kotlin/com/neoutils/neoregex/App.desktop.kt
@@ -47,7 +47,6 @@ import com.neoutils.neoregex.core.datasource.model.Preferences
 import com.neoutils.neoregex.core.datasource.remember.rememberWindowState
 import com.neoutils.neoregex.core.designsystem.theme.NeoTheme
 import com.neoutils.neoregex.core.designsystem.theme.NeoTheme.dimensions
-import com.neoutils.neoregex.core.dispatcher.di.dispatcherModule
 import com.neoutils.neoregex.core.manager.di.managerModule
 import com.neoutils.neoregex.core.repository.di.repositoryModule
 import com.neoutils.neoregex.core.sharedui.component.*
@@ -66,7 +65,6 @@ fun ApplicationScope.DesktopApp() = WithKoin(
     dataSourceModule,
     databaseModule,
     repositoryModule,
-    dispatcherModule,
     matcherModule,
     validatorModule,
     savedModule,
