--- core/dispatcher/src/androidMain/kotlin/com/neoutils/neoregex/core/dispatcher/di/NavigationModule.android.kt
@@ -1,28 +1,0 @@
-/*
- * NeoRegex.
- *
- * Copyright (C) 2024 Irineu A. Silva.
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <https://www.gnu.org/licenses/>.
- */
-
-package com.neoutils.neoregex.core.dispatcher.di
-
-import com.neoutils.neoregex.core.dispatcher.navigator.NavigationManager
-import com.neoutils.neoregex.core.dispatcher.navigator.NavigationManagerImpl
-import org.koin.dsl.bind
-import org.koin.dsl.module
-
-actual val dispatcherModule = module {
-    single { NavigationManagerImpl() } bind NavigationManager::class
-}
