+++ core/manager/src/commonMain/kotlin/com/neoutils/neoregex/core/manager/model/Opened.kt
@@ -1,7 +1,7 @@
 /*
  * NeoRegex.
  *
- * Copyright (C) 2025 <AUTHOR>.
+ * Copyright (C) 2025 Irineu A. Silva.
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
